chore(objectmanager): fix race condition using networkmock

+++ pkg/objectmanager/objectmanager_test.go
@@ -760,7 +760,7 @@ func Test_manager_Put(t *testing.T) {
 				t.Cleanup(func() {
 					// TODO properly test this, this should be 1 but due to
 					// go m.announceObject it's not
-					assert.Equal(t, int32(0), m.SendCalled)
+					assert.Equal(t, 0, m.SendCalled())
 				})
 				return m
 			},
@@ -851,7 +851,7 @@ func Test_manager_Put(t *testing.T) {
 				t.Cleanup(func() {
 					// TODO properly test this, this should be 1 but due to
 					// go m.announceObject it's not
-					assert.Equal(t, int32(0), m.SendCalled)
+					assert.Equal(t, 0, m.SendCalled())
 				})
 				return m
 			},
